requirements.txt
fastapi==0.95.0
tensorflow==2.13.0
pillow==10.0.0
uvicorn==0.22.0
numpy==1.24.3
pytest==7.4.0

